requirements.txt
shap==0.39.0
streamlit==0.82.0
matplotlib==3.2.2
numpy==1.20.2
joblib==0.15.1
pandas==1.0.4
Pillow==8.2.0
scikit-learn==0.24.2
scipy==1.6.3
Flask==1.1.2